Rename modules
* opencv -> detection
* detection -> engine
