+++ detection/.gitignore
@@ -1,1 +1,2 @@
 /build
+/src/main/cpp/opencv
