+++ detection/src/main/AndroidManifest.xml
@@ -1,18 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<!--
- * Copyright (C) 2020 Nain57
- *
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU General Public License
- * as published by the Free Software Foundation; either version 2
- * of the License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program; If not, see <http://www.gnu.org/licenses/>.
--->
-<manifest package="com.buzbuz.smartautoclicker.detection"/>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="com.buzbuz.smartautoclicker.detection">
+
+</manifest>
