+++ detection/src/main/cpp/detector.hpp
