--- detection/src/test/java/com/buzbuz/smartautoclicker/detection/utils/Matchers.kt
@@ -1,25 +1,0 @@
-/*
- * Copyright (C) 2021 Nain57
- *
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU General Public License
- * as published by the Free Software Foundation; either version 2
- * of the License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program; If not, see <http://www.gnu.org/licenses/>.
- */
-package com.buzbuz.smartautoclicker.detection.utils
-
-import org.mockito.Mockito
-
-/**
- * Argument matcher for a nonnull kotlin function argument matching any nonnull value.
- * @param T the type of the argument.
- */
-fun <T> anyNotNull(): T = Mockito.any()
