--- opencv/.gitignore
@@ -1,2 +1,0 @@
-/build
-/src/main/cpp/opencv
