+++ settings.gradle
@@ -1,7 +1,7 @@
 include ':extensions'
-include ':detection'
+include ':engine'
 include ':database'
 include ':base-ui'
 rootProject.name='Smart Auto Clicker'
 include ':smartautoclicker'
-include ':opencv'
+include ':detection'
